fix offer component, disable HOC
